config: factor yaml parsing out into a separate silkit_yaml
