+++ CMakeLists.txt
@@ -26,6 +26,7 @@ option(SILKIT_USE_SYSTEM_LIBRARIES "Use the libraries installed on the system fo
 option(SILKIT_BUILD_REPRODUCIBLE "Creates a reproducible build by ommiting timestamps/unique build ids" ON)
 option(SILKIT_BUILD_LINUX_PACKAGE "Creates SIL Kit builds suitable for package managers in Linux Distributions (.deb)" OFF)
 option(SILKIT_BUILD_LTO "Build with link time optimizations" OFF)
+option(SILKIT_INSTALL_YAML "Install the SilKit::Yaml header-only auxiliary YAML library (silkit_yaml; unstable API/ABI)" OFF)
 
 if(SILKIT_BUILD_LINUX_PACKAGE)
